+++ src/test/java/com/devcourse/springjpaboard/user/controller/UserControllerTest.java
@@ -1,11 +1,24 @@
 package com.devcourse.springjpaboard.user.controller;
 
+import static com.devcourse.springjpaboard.core.exception.ExceptionMessage.NOT_FOUND_USER;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.doThrow;
+import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
+import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.documentationConfiguration;
+import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
+import static org.springframework.restdocs.payload.PayloadDocumentation.requestFields;
+import static org.springframework.restdocs.payload.PayloadDocumentation.responseFields;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
 import com.devcourse.springjpaboard.application.user.controller.UserController;
-import com.devcourse.springjpaboard.core.exception.NotFoundException;
 import com.devcourse.springjpaboard.application.user.controller.dto.CreateUserRequest;
 import com.devcourse.springjpaboard.application.user.controller.dto.UserResponse;
-import com.devcourse.springjpaboard.user.controller.stub.UserStubs;
 import com.devcourse.springjpaboard.application.user.service.UserService;
+import com.devcourse.springjpaboard.core.exception.GlobalExceptionHandler;
+import com.devcourse.springjpaboard.core.exception.NotFoundException;
+import com.devcourse.springjpaboard.user.controller.stub.UserStubs;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -24,113 +37,105 @@ import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
-import static com.devcourse.springjpaboard.core.exception.ExceptionMessage.NOT_FOUND_USER;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.*;
-import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
-import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.documentationConfiguration;
-import static org.springframework.restdocs.payload.PayloadDocumentation.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
 @ExtendWith({MockitoExtension.class, RestDocumentationExtension.class})
 class UserControllerTest {
 
-    @InjectMocks
-    private UserController userController;
+  @InjectMocks
+  private UserController userController;
 
-    @Mock
-    private UserService userService;
+  @Mock
+  private UserService userService;
 
-    private MockMvc mockMvc;
+  private MockMvc mockMvc;
 
-    private final ObjectMapper mapper = new ObjectMapper();
+  private final ObjectMapper mapper = new ObjectMapper();
 
-    @BeforeEach
-    public void init(RestDocumentationContextProvider restDocumentationContextProvider) {
-        mockMvc = MockMvcBuilders.standaloneSetup(userController)
-                .apply(documentationConfiguration(restDocumentationContextProvider))
-                .build();
-    }
+  @BeforeEach
+  public void init(RestDocumentationContextProvider restDocumentationContextProvider) {
+    mockMvc = MockMvcBuilders.standaloneSetup(userController)
+        .apply(documentationConfiguration(restDocumentationContextProvider))
+        .setControllerAdvice(new GlobalExceptionHandler())
+        .build();
+  }
 
-    @Test
-    @DisplayName("회원가입 테스트")
-    void createUserTest() throws Exception {
-        // given
-        CreateUserRequest request = UserStubs.createUserRequest();
-        UserResponse response = UserStubs.userResponse();
+  @Test
+  @DisplayName("회원가입 테스트")
+  void createUserTest() throws Exception {
+    // given
+    CreateUserRequest request = UserStubs.createUserRequest();
+    UserResponse response = UserStubs.userResponse();
 
-        doReturn(response)
-                .when(userService)
-                .save(any(CreateUserRequest.class));
+    doReturn(response)
+        .when(userService)
+        .save(any(CreateUserRequest.class));
 
-        // when
-        ResultActions resultActions = mockMvc.perform(
-                MockMvcRequestBuilders.post("/users")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(mapper.writeValueAsString(request))
-        );
+    // when
+    ResultActions resultActions = mockMvc.perform(
+        MockMvcRequestBuilders.post("/users")
+            .contentType(MediaType.APPLICATION_JSON)
+            .content(mapper.writeValueAsString(request))
+    );
 
-        // then
-        resultActions.andExpect(status().isOk())
-                .andDo(document("user-save",
-                        requestFields(
-                                fieldWithPath("name").type(JsonFieldType.STRING).description("String"),
-                                fieldWithPath("age").type(JsonFieldType.NUMBER).description("int"),
-                                fieldWithPath("hobby").type(JsonFieldType.STRING).description("String")
-                        ),
-                        responseFields(
-                                fieldWithPath("statusCode").type(JsonFieldType.NUMBER).description("상태코드"),
-                                fieldWithPath("data.id").type(JsonFieldType.NUMBER).description("아이디"),
-                                fieldWithPath("data.name").type(JsonFieldType.STRING).description("이름"),
-                                fieldWithPath("data.age").type(JsonFieldType.NUMBER).description("나이"),
-                                fieldWithPath("data.hobby").type(JsonFieldType.STRING).description("취미"),
-                                fieldWithPath("serverDateTime").type(JsonFieldType.STRING).description("응답시간")
-                        )
-                ));
-    }
+    // then
+    resultActions.andExpect(status().isOk())
+        .andDo(document("user-save",
+            requestFields(
+                fieldWithPath("name").type(JsonFieldType.STRING).description("String"),
+                fieldWithPath("age").type(JsonFieldType.NUMBER).description("int"),
+                fieldWithPath("hobby").type(JsonFieldType.STRING).description("String")
+            ),
+            responseFields(
+                fieldWithPath("statusCode").type(JsonFieldType.NUMBER).description("상태코드"),
+                fieldWithPath("data.id").type(JsonFieldType.NUMBER).description("아이디"),
+                fieldWithPath("data.name").type(JsonFieldType.STRING).description("이름"),
+                fieldWithPath("data.age").type(JsonFieldType.NUMBER).description("나이"),
+                fieldWithPath("data.hobby").type(JsonFieldType.STRING).description("취미"),
+                fieldWithPath("serverDateTime").type(JsonFieldType.STRING).description("응답시간")
+            )
+        ));
+  }
 
-    @Test
-    @DisplayName("존재하는 회원 ID로 조회 테스트")
-    void getUserByIdTest() throws Exception {
-        // given
-        Long request = 1L;
-        UserResponse response = UserStubs.userResponse();
+  @Test
+  @DisplayName("존재하는 회원 ID로 조회 테스트")
+  void getUserByIdTest() throws Exception {
+    // given
+    Long request = 1L;
+    UserResponse response = UserStubs.userResponse();
 
-        doReturn(response).when(userService).findById(request);
-        // when
-        ResultActions resultActions = mockMvc.perform(
-                MockMvcRequestBuilders.get("/users/1")
-        );
-        // then
-        resultActions.andExpect(status().isOk())
-                .andDo(document("user-find-by-id",
-                        responseFields(
-                                fieldWithPath("statusCode").type(JsonFieldType.NUMBER).description("상태코드"),
-                                fieldWithPath("data.id").type(JsonFieldType.NUMBER).description("아이디"),
-                                fieldWithPath("data.name").type(JsonFieldType.STRING).description("이름"),
-                                fieldWithPath("data.age").type(JsonFieldType.NUMBER).description("나이"),
-                                fieldWithPath("data.hobby").type(JsonFieldType.STRING).description("취미"),
-                                fieldWithPath("serverDateTime").type(JsonFieldType.STRING).description("응답시간")
-                        )
-                ));
-    }
+    doReturn(response).when(userService).findById(request);
+    // when
+    ResultActions resultActions = mockMvc.perform(
+        MockMvcRequestBuilders.get("/users/1")
+    );
+    // then
+    resultActions.andExpect(status().isOk())
+        .andDo(document("user-find-by-id",
+            responseFields(
+                fieldWithPath("statusCode").type(JsonFieldType.NUMBER).description("상태코드"),
+                fieldWithPath("data.id").type(JsonFieldType.NUMBER).description("아이디"),
+                fieldWithPath("data.name").type(JsonFieldType.STRING).description("이름"),
+                fieldWithPath("data.age").type(JsonFieldType.NUMBER).description("나이"),
+                fieldWithPath("data.hobby").type(JsonFieldType.STRING).description("취미"),
+                fieldWithPath("serverDateTime").type(JsonFieldType.STRING).description("응답시간")
+            )
+        ));
+  }
 
-    @Test
-    @DisplayName("존재하지 않는 회원 ID로 조회시 예외 발생 테스트")
-    void getUserByIdFailTest() throws Exception {
-        // given
-        Long request = 2L;
-        doThrow(new NotFoundException(NOT_FOUND_USER)).when(userService).findById(request);
+  @Test
+  @DisplayName("존재하지 않는 회원 ID로 조회시 예외 발생 테스트")
+  void getUserByIdFailTest() throws Exception {
+    // given
+    Long request = 2L;
+    doThrow(new NotFoundException(NOT_FOUND_USER)).when(userService).findById(request);
 
-        // when
-        ResultActions resultActions = mockMvc.perform(
-                MockMvcRequestBuilders.get("/users/2")
-        );
-        // then
-        MvcResult mvcResult = resultActions
-                .andExpect(status().is4xxClientError())
-                .andExpect(jsonPath("data").exists())
-                .andReturn();
-    }
+    // when
+    ResultActions resultActions = mockMvc.perform(
+        MockMvcRequestBuilders.get("/users/2")
+    );
+    // then
+    MvcResult mvcResult = resultActions
+        .andExpect(status().is4xxClientError())
+        .andExpect(jsonPath("data").exists())
+        .andReturn();
+  }
 }
